requirement.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-Migrate==1.3.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-WTF==0.10.3
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
SQLAlchemy==0.9.8
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.6
alembic==0.7.2
argparse==1.2.1
decorator==3.4.0
itsdangerous==0.24
pbr==0.10.7
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.14
wsgiref==0.1.2
